Rewrite for runc
As a design decision, runc does not like mounting anything.

Instead of mounting files, we'll set up a couple of things for our
containers:

1. a web server with a `run-script` containing $VALIDATION_SCRIPT
2. a tarball containing per configuration

We'll tell our containers to start by downloading the tarball
and extracting it.

Then we'll download and run the run-script.

To improve performance, we start neo4j in the background and
pull the validation image in the background.

We don't run configuration validation until neo4j is ready,
and we wait for the validation image to be available before
starting the validators. We expect that the image will
arrive before neo4j is ready.